hour rate could be only integer

+++ src/main/java/com/stormpath/tutorial/controller/jsonrequest/HourRate.java
@@ -4,13 +4,13 @@ package com.stormpath.tutorial.controller.jsonrequest;
  * Created by tomasz.lelek on 21/12/15.
  */
 public class HourRate {
-    public Double hourRate;
+    public Integer hourRate;
 
-    public Double getHourRate() {
+    public Integer getHourRate() {
         return hourRate;
     }
 
-    public void setHourRate(Double hourRate) {
+    public void setHourRate(Integer hourRate) {
         this.hourRate = hourRate;
     }
 }
